Patch from Seb Wills:
Hi Richard,

Trying to add http://www.nature.com/nrn/journal/v5/n2/abs/nrn1327_fs.html
results in "Couldn't fetch the citation details from the Nature web site."

I had a look at nature.pl in the svn repository, and the problem seems to be that the URL is a mixture of what your comments call "Old style" and new style Nature URLs.

Specifically, it does not have "/cgi" in it, so does not match the "Old style" pattern, but the last element of the URL does have a "_fs" tagged onto the end of the filename which the new style pattern fails to strip off. The RIS file is in fact at http://www.nature.com/nrn/journal/v5/n2/ris/nrn1327.ris (without the _fs).

I think that you just need to insert one character (a "_" into the final character spec) to fix this:

54c54
< } elsif ($url =~ m{www.nature.com/([^/]+)/journal/v([0-9]+)/n([0-9]+)/[^/]+/([^/.]+)}) {
---
} elsif ($url =~ m{www.nature.com/([^/]+)/journal/v([0-9]+)/n([0-9]+)/[^/]+/([^/._]+)}) {

cheers,

seb



git-svn-id: http://svn.citeulike.org/svn/plugins@32 eb686afc-39fa-0310-a79f-9b3c69cefa1c